Marked final obsolete functions and split ColorHelper with ColorExtensions
